移植 26.1.2 NeoForge

+++ src/main/java/org/mesdag/particlestorm/data/molang/compiler/function/round/LerpFunction.java
@@ -5,11 +5,14 @@ import org.mesdag.particlestorm.api.MolangInstance;
 import org.mesdag.particlestorm.data.molang.compiler.MathValue;
 import org.mesdag.particlestorm.data.molang.compiler.function.MathFunction;
 
-/// [MathFunction] value supplier
-///
-/// **Contract:**
-///
-/// Returns the first value plus the difference between the first and second input values multiplied by the third input value
+/**
+ * {@link MathFunction} value supplier
+ *
+ * <p>
+ * <b>Contract:</b>
+ * <br>
+ * Returns the first value plus the difference between the first and second input values multiplied by the third input value
+ */
 public final class LerpFunction extends MathFunction {
     private final MathValue min;
     private final MathValue max;
@@ -29,7 +32,7 @@ public final class LerpFunction extends MathFunction {
     }
 
     @Override
-    public float compute(MolangInstance instance) {
+    public double compute(MolangInstance instance) {
         return Mth.lerp(this.delta.get(instance), this.min.get(instance), this.max.get(instance));
     }
